// frontend-legacy/src/app/shared/enums/chart-type.enum.ts
export enum ChartTypeEnum {
  LINE_CHART = 'LINE_CHART',
  AREA_LINE_CHART = 'AREA_CHART',
  PIE_CHART = 'PIE_CHART',
  BAR_CHART = 'VERTICAL_BAR_CHART',
  BAR_HORIZONTAL_CHART = 'HORIZONTAL_BAR_CHART',
  TAG_CLOUD_CHART = 'TAG_CLOUD_CHART',
  TABLE_CHART = 'TABLE_CHART',
  GAUGE_CHART = 'GAUGE_CHART',
  GOAL_CHART = 'GOAL_CHART',
  METRIC_CHART = 'METRIC_CHART',
  HEATMAP_CHART = 'HEATMAP_CHART',
  MARKER_CHART = 'COORDINATE_MAP_CHART',
  LIST_CHART = 'LIST_CHART',
  TEXT_CHART = 'TEXT_CHART',
  VERTICAL_BAR_CHART = 'VERTICAL_BAR_CHART'
}